+++ src/Orchestrator/HackAlgorithm/MoneyHackAlgorithm.ts
@@ -7,35 +7,42 @@ export function MoneyHackAlgorithm(ns: NS, currentHack: Hack[], hackedHost: Hack
     //DEBUG && ns.print("Calculating hacks")
     let potentialHack: Hack[] = []
 
-    for (let i = 0; i < hackedHost.length; i++) {
-        if (hackedHost[i].maxMoney === 0) {
+    for (const host of hackedHost) {
+        if (host.maxMoney === 0) {
             continue
         }
 
-        if (currentHack.find(h => h.host == hackedHost[i].name)) {
+        if (currentHack.find(h => h.host == host.name)) {
             continue
         }
 
         // Quick hack
-        const hostCurMoney: number = ns.getServerMoneyAvailable(hackedHost[i].name)
-        const hostCurSecurity: number = ns.getServerSecurityLevel(hackedHost[i].name)
+        const hostCurMoney: number = ns.getServerMoneyAvailable(host.name)
+        const hostCurSecurity: number = ns.getServerSecurityLevel(host.name)
         const maxHackAmount: number = hostCurMoney * MONEY_HACKING_TARGET_PERCENT
-        const hackThreads: number = Math.ceil(ns.hackAnalyzeThreads(hackedHost[i].name, maxHackAmount))
+        const hackThreads: number = Math.ceil(ns.hackAnalyzeThreads(host.name, maxHackAmount))
+        const finalHackThreads: number = Math.min(hackThreads, availableThreads)
         const hackPercentage: number = hackThreads/availableThreads > 1 ? 1 : hackThreads/availableThreads
         const hackAmount: number = maxHackAmount * hackPercentage
-        const baseHackChance: number = ((1.75 * ns.getHackingLevel()) - hackedHost[i].hackingRequired)/(1.75 * ns.getHackingLevel())
+        const baseHackChance: number = ((1.75 * ns.getHackingLevel()) - host.hackingRequired)/(1.75 * ns.getHackingLevel())
         const hackChance: number = (100-hostCurSecurity)/100*baseHackChance
+        const hackPerThread: number = ns.hackAnalyze(host.name)
+
+        // We skip those that are not high enough
+        // if (hostCurMoney/host.maxMoney < 0.75) {
+        //     continue
+        // }
 
         // We need to ensure that it return a valid number of thread for the hack
-        if (Number.isFinite(hackThreads) && hackThreads > 0) {
+        if (Number.isFinite(finalHackThreads) && hackThreads > 0) {
             potentialHack.push(new Hack(
-                hackedHost[i].name,
-                hackedHost[i].hackTime,
+                host.name,
+                host.hackTime,
                 hackAmount,
-                hackThreads,
+                finalHackThreads,
                 0,
                 0,
-                hackAmount / hackedHost[i].hackTime*hackChance,
+                (hackAmount*hackPerThread) / host.hackTime * hackChance,
                 HackType.moneyHack,
                 hackChance
             ))